Part 4 continuation
Completed Part 4 (2,3,4) Im still just a bit confused about the third feature if I added a right one. Also, when plotting I realized that my data was getting compressed when having to analyse GNI there I searched up on the internet and fount out about. plt.sxcale("log") This will change our x axis to a logarithmic scale instead of a linear scale to help see the pattern better. When plotting it, the result was very compresed. So with log, it will be less compressed by spreading out the smaller values and compressing the larger ones. ---------here: https://www.geeksforgeeks.org/python/how-to-plot-logarithmic-axes-in-matplotlib/ -----------(I<m still not sure if I am allowed to use this, but i didn't know what else to do)